+++ src/components/bouquets/BouquetList.vue
@@ -5,7 +5,7 @@ import { computed, onMounted } from 'vue'
 import { useLoading } from 'vue-loading-overlay'
 import { useRoute, useRouter, type LocationQueryRaw } from 'vue-router'
 
-import BouquetCard from '@/components/bouquets/BouquetCard.vue'
+import BouquetRow from '@/components/bouquets/BouquetRow.vue'
 import { NoOptionSelected } from '@/model/theme'
 import type { Topic } from '@/model/topic'
 import { useTopicStore } from '@/store/TopicStore'
@@ -174,7 +174,7 @@ defineExpose({
   <div class="bouquets-list-container fr-container fr-mb-4w border-top">
     <div class="fr-grid-row fr-grid-row--gutters fr-mb-1w">
       <div class="fr-col-12">
-        <BouquetCard
+        <BouquetRow
           v-for="bouquet in bouquets"
           :key="bouquet.id"
           :bouquet="bouquet"
